refactor(core): extract windowed file transcription
Move stream_eligible, transcribe_stream_file, and
transcribe_stream_mono into engine/file_stream.rs. The request
pipeline stays in transcribe.rs; public Engine methods are
unchanged.
